+++ cmd/livesim2/app/config.go
@@ -18,6 +18,7 @@ import (
 	"github.com/knadh/koanf/providers/posflag"
 	"github.com/knadh/koanf/providers/structs"
 
+	"github.com/Dash-Industry-Forum/livesim2/pkg/drm"
 	"github.com/Dash-Industry-Forum/livesim2/pkg/logging"
 	"github.com/spf13/pflag"
 )
@@ -60,7 +61,9 @@ type ServerConfig struct {
 	Host string `json:"host"`
 	// PlayURL is a URL template to play asset including player and pattern %s to be replaced by MPD URL
 	// For autoplay, start the player muted.
-	PlayURL string `json:"playurl"`
+	PlayURL    string         `json:"playurl"`
+	DrmCfgFile string         `json:"drmcfgfile"`
+	DrmCfg     *drm.DrmConfig `json:"drmcfg"`
 }
 
 var DefaultConfig = ServerConfig{
@@ -126,6 +129,8 @@ func LoadConfig(args []string, cwd string) (*ServerConfig, error) {
 	f.String("scheme", k.String("scheme"), "scheme used in Location and BaseURL elements. If empty, it is attempted to be auto-detected")
 	f.String("host", k.String("host"), "host (and possible prefix) used in MPD elements. Overrides auto-detected full scheme://host")
 	f.String("playurl", k.String("playurl"), "URL template to play mpd. %s will be replaced by MPD URL")
+	f.String("drmcfgfile", k.String("drmcfgfile"), "DRM config file path")
+
 	if err := f.Parse(args[1:]); err != nil {
 		return nil, fmt.Errorf("command line parse: %w", err)
 	}
@@ -158,15 +163,9 @@ func LoadConfig(args []string, cwd string) (*ServerConfig, error) {
 	}
 
 	// Make vodPath absolute in case it is not already
-	vodRoot := k.String("vodroot")
-	if vodRoot != "" && !path.IsAbs(vodRoot) {
-		vodRoot = path.Join(cwd, vodRoot)
-		err = k.Load(confmap.Provider(map[string]any{
-			"vodroot": vodRoot,
-		}, "."), nil)
-		if err != nil {
-			return nil, err
-		}
+	vodRoot, err := makeAbsolutePath(k, "vodroot", cwd)
+	if err != nil {
+		return nil, fmt.Errorf("make vodroot absolute: %w", err)
 	}
 	// Update repDataRoot to consistent value including absolute path
 	repDataRoot := k.String("repdataroot")
@@ -215,6 +214,12 @@ func LoadConfig(args []string, cwd string) (*ServerConfig, error) {
 		}
 	}
 
+	// Make drmconfig absolute in case it is not already
+	_, err = makeAbsolutePath(k, "drmconfig", cwd)
+	if err != nil {
+		return nil, fmt.Errorf("make drmconfig absolute: %w", err)
+	}
+
 	// Unmarshal into cfg
 	var cfg ServerConfig
 	if err := k.Unmarshal("", &cfg); err != nil {
@@ -222,6 +227,20 @@ func LoadConfig(args []string, cwd string) (*ServerConfig, error) {
 	}
 
 	return &cfg, nil
+}
+
+func makeAbsolutePath(k *koanf.Koanf, key, cwd string) (string, error) {
+	absPath := k.String(key)
+	if absPath != "" && !path.IsAbs(absPath) {
+		absPath = path.Join(cwd, absPath)
+		err := k.Load(confmap.Provider(map[string]any{
+			key: absPath,
+		}, "."), nil)
+		if err != nil {
+			return "", err
+		}
+	}
+	return absPath, nil
 }
 
 func checkTLSParams(k *koanf.Koanf) error {