+++ cmd/livesim2/app/livesegment_test.go
@@ -16,6 +16,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/Dash-Industry-Forum/livesim2/pkg/drm"
 	"github.com/Dash-Industry-Forum/livesim2/pkg/logging"
 	"github.com/Eyevinn/mp4ff/bits"
 	"github.com/Eyevinn/mp4ff/mp4"
@@ -29,6 +30,7 @@ func TestLiveSegment(t *testing.T) {
 	logger := slog.Default()
 	err := am.discoverAssets(logger)
 	require.NoError(t, err)
+	log := slog.Default()
 
 	cases := []struct {
 		asset           string
@@ -59,6 +61,7 @@ func TestLiveSegment(t *testing.T) {
 			mediaTimescale:  48000,
 		},
 	}
+	var drmCfg *drm.DrmConfig = nil
 	for _, tc := range cases {
 		t.Run(tc.asset, func(t *testing.T) {
 			for _, mpdType := range []string{"Number", "TimelineTime"} {
@@ -75,11 +78,11 @@ func TestLiveSegment(t *testing.T) {
 				}
 				nowMS := 100_000
 				rr := httptest.NewRecorder()
-				wroteInit, err := writeInitSegment(rr, cfg, asset, "2/init.mp4")
+				wroteInit, err := writeInitSegment(log, rr, cfg, drmCfg, asset, "2/init.mp4")
 				require.False(t, wroteInit)
 				require.NoError(t, err)
 				rr = httptest.NewRecorder()
-				wroteInit, err = writeInitSegment(rr, cfg, asset, tc.initialization)
+				wroteInit, err = writeInitSegment(log, rr, cfg, drmCfg, asset, tc.initialization)
 				require.True(t, wroteInit)
 				require.NoError(t, err)
 				require.Equal(t, http.StatusOK, rr.Code)
@@ -98,7 +101,7 @@ func TestLiveSegment(t *testing.T) {
 				default: // "TimelineTime":
 					media = strings.Replace(media, "$NrOrTime$", fmt.Sprintf("%d", mediaTime), -1)
 				}
-				so, err := genLiveSegment(vodFS, asset, cfg, media, nowMS, false /*isLast */)
+				so, err := genLiveSegment(log, vodFS, asset, cfg, media, nowMS, false /*isLast */)
 				require.NoError(t, err)
 				require.Equal(t, tc.segmentMimeType, so.meta.rep.SegmentType())
 				seg := so.seg
@@ -113,8 +116,8 @@ func TestLiveSegment(t *testing.T) {
 func TestAc3Timing(t *testing.T) {
 	vodFS := os.DirFS("testdata/assets")
 	am := newAssetMgr(vodFS, "", false)
-	logger := slog.Default()
-	err := am.discoverAssets(logger)
+	log := slog.Default()
+	err := am.discoverAssets(log)
 	require.NoError(t, err)
 
 	asset, ok := am.findAsset("bbb_hevc_ac3_8s")
@@ -124,7 +127,7 @@ func TestAc3Timing(t *testing.T) {
 	for sNr := 0; sNr <= 5; sNr++ {
 		media := "audio_$NrOrTime$.m4s"
 		media = strings.Replace(media, "$NrOrTime$", fmt.Sprintf("%d", sNr), -1)
-		so, err := genLiveSegment(vodFS, asset, cfg, media, nowMS, false /* isLast */)
+		so, err := genLiveSegment(log, vodFS, asset, cfg, media, nowMS, false /* isLast */)
 		require.NoError(t, err)
 		bmdt := int(so.seg.Fragments[0].Moof.Traf.Tfdt.BaseMediaDecodeTime())
 		overShoot := bmdt - (2 * sNr * 48000)
@@ -136,8 +139,8 @@ func TestAc3Timing(t *testing.T) {
 func TestCheckAudioSegmentTimeAddressing(t *testing.T) {
 	vodFS := os.DirFS("testdata/assets")
 	am := newAssetMgr(vodFS, "", false)
-	logger := slog.Default()
-	err := am.discoverAssets(logger)
+	log := slog.Default()
+	err := am.discoverAssets(log)
 	require.NoError(t, err)
 
 	cases := []struct {
@@ -178,7 +181,7 @@ func TestCheckAudioSegmentTimeAddressing(t *testing.T) {
 				default:
 					segMedia = strings.Replace(c.media, "$NrOrTime$", fmt.Sprintf("%d", mediaTime), -1)
 				}
-				so, err := genLiveSegment(vodFS, asset, cfg, segMedia, c.nowMS, false /* isLast */)
+				so, err := genLiveSegment(log, vodFS, asset, cfg, segMedia, c.nowMS, false /* isLast */)
 				require.NoError(t, err)
 				trun := so.seg.Fragments[0].Moof.Traf.Trun
 				nrSamples := c.nrSamplesMod[nr%len(c.nrSamplesMod)]
@@ -192,8 +195,8 @@ func TestCheckAudioSegmentTimeAddressing(t *testing.T) {
 func TestLiveThumbSegment(t *testing.T) {
 	vodFS := os.DirFS("testdata/assets")
 	am := newAssetMgr(vodFS, "", false)
-	logger := slog.Default()
-	err := am.discoverAssets(logger)
+	log := slog.Default()
+	err := am.discoverAssets(log)
 	require.NoError(t, err)
 
 	cases := []struct {
@@ -231,7 +234,7 @@ func TestLiveThumbSegment(t *testing.T) {
 			media := tc.media
 			// Always number, even if MPD is timelinetime
 			media = strings.Replace(media, "$NrOrTime$", fmt.Sprintf("%d", tc.reqNr), -1)
-			so, err := genLiveSegment(vodFS, asset, cfg, media, nowMS, false /* isLast */)
+			so, err := genLiveSegment(log, vodFS, asset, cfg, media, nowMS, false /* isLast */)
 			require.NoError(t, err)
 			origNr := tc.reqNr%tc.nrSegs + 1 // one-based
 			require.Equal(t, tc.segmentMimeType, so.meta.rep.SegmentType())
@@ -245,8 +248,8 @@ func TestLiveThumbSegment(t *testing.T) {
 func TestWriteChunkedSegment(t *testing.T) {
 	vodFS := os.DirFS("testdata/assets")
 	am := newAssetMgr(vodFS, "", false)
-	logger := slog.Default()
-	err := am.discoverAssets(logger)
+	log := slog.Default()
+	err := am.discoverAssets(log)
 	require.NoError(t, err)
 	cfg := NewResponseConfig()
 	cfg.AvailabilityTimeCompleteFlag = false
@@ -274,7 +277,7 @@ func TestWriteChunkedSegment(t *testing.T) {
 		rr := httptest.NewRecorder()
 		segmentPart := strings.Replace(tc.media, "$NrOrTime$", "10", 1)
 		mediaTime := 80 * tc.mediaTimescale
-		err := writeChunkedSegment(context.Background(), rr, cfg, vodFS, asset, segmentPart, nowMS, false /* isLast */)
+		err := writeChunkedSegment(context.Background(), log, rr, cfg, nil, vodFS, asset, segmentPart, nowMS, false /* isLast */)
 		require.NoError(t, err)
 		seg := rr.Body.Bytes()
 		sr := bits.NewFixedSliceReader(seg)
@@ -389,8 +392,8 @@ func TestTTMLTimeShifts(t *testing.T) {
 func TestStartNumber(t *testing.T) {
 	vodFS := os.DirFS("testdata/assets")
 	am := newAssetMgr(vodFS, "", false)
-	logger := slog.Default()
-	err := am.discoverAssets(logger)
+	log := slog.Default()
+	err := am.discoverAssets(log)
 	require.NoError(t, err)
 	err = logging.InitSlog("debug", "discard")
 	require.NoError(t, err)
@@ -438,7 +441,7 @@ func TestStartNumber(t *testing.T) {
 		cfg := NewResponseConfig()
 		cfg.StartNr = Ptr(tc.startNr)
 		media := strings.Replace(tc.media, "$NrOrTime$", fmt.Sprintf("%d", (tc.requestNr)), 1)
-		so, err := genLiveSegment(vodFS, asset, cfg, media, tc.nowMS, false /* isLast */)
+		so, err := genLiveSegment(log, vodFS, asset, cfg, media, tc.nowMS, false /* isLast */)
 		if tc.expectedErr != "" {
 			require.Error(t, err)
 			require.Contains(t, err.Error(), tc.expectedErr)
@@ -457,8 +460,8 @@ func TestStartNumber(t *testing.T) {
 func TestLLSegmentAvailability(t *testing.T) {
 	vodFS := os.DirFS("testdata/assets")
 	am := newAssetMgr(vodFS, "", false)
-	logger := slog.Default()
-	err := am.discoverAssets(logger)
+	log := slog.Default()
+	err := am.discoverAssets(log)
 	require.NoError(t, err)
 	err = logging.InitSlog("error", "discard")
 	require.NoError(t, err)
@@ -576,7 +579,7 @@ func TestLLSegmentAvailability(t *testing.T) {
 			cfg.StartNr = Ptr(tc.startNr)
 		}
 		media := strings.Replace(tc.media, "$NrOrTime$", fmt.Sprintf("%d", (tc.requestMedia)), 1)
-		so, err := genLiveSegment(vodFS, asset, cfg, media, tc.nowMS, false /* isLast */)
+		so, err := genLiveSegment(log, vodFS, asset, cfg, media, tc.nowMS, false /* isLast */)
 		if tc.expectedErr != "" {
 			require.Error(t, err)
 			require.Contains(t, err.Error(), tc.expectedErr)
@@ -697,7 +700,8 @@ func TestSegmentStatusCodeResponse(t *testing.T) {
 			cfg.SegStatusCodes = tc.ss
 			media := strings.Replace(tc.media, "$NrOrTime$", fmt.Sprintf("%d", tc.nrOrTime), -1)
 			rr := httptest.NewRecorder()
-			code, err := writeSegment(context.TODO(), rr, slog.Default(), cfg, vodFS, asset, media, tc.nowMS, nil, false /* isLast */)
+			code, err := writeSegment(context.TODO(), rr, slog.Default(), cfg, nil,
+				vodFS, asset, media, tc.nowMS, nil, false /* isLast */)
 			require.NoError(t, err)
 			require.Equal(t, tc.expCode, code)
 		})
@@ -705,6 +709,7 @@ func TestSegmentStatusCodeResponse(t *testing.T) {
 }
 
 func TestMehdBoxRemovedFromInitSegment(t *testing.T) {
+	var drmCfg *drm.DrmConfig = nil
 	vodFS := os.DirFS("testdata/assets")
 	am := newAssetMgr(vodFS, "", false)
 	logger := slog.Default()
@@ -714,7 +719,7 @@ func TestMehdBoxRemovedFromInitSegment(t *testing.T) {
 	require.True(t, ok)
 	cfg := NewResponseConfig()
 	initV300 := "V300/init.mp4"
-	match, err := matchInit(initV300, cfg, asset)
+	match, err := matchInit(initV300, cfg, drmCfg, asset)
 	require.NoError(t, err)
 	sr := bits.NewFixedSliceReader(match.init)
 	mp4File, err := mp4.DecodeFileSR(sr)